Got UI to render, need to update the menus
